Integer для user_id

+++ pre-revolutionary-tinder/src/main/java/ru/ligaintenship/prerevolutionarytinder/domain/UserCreator.java
@@ -11,7 +11,7 @@ public class UserCreator {
 
     public void create(User resource) {
         String sql = "insert into tinder.tinder_users (user_id, sex, name, story, looking_for) values (%d, '%s', '%s', '%s', '%s')"
-                .formatted(resource.getUser_id().intValue(), resource.getSex(), resource.getName(), resource.getStory(), resource.getLooking_for());
+                .formatted(resource.getUser_id(), resource.getSex(), resource.getName(), resource.getStory(), resource.getLooking_for());
         int responseCode = provider.putData(sql);
         System.out.println(responseCode);
     }
